refactor: 타이머 폴더 구조 변경
