+++ src/components/stopwatch/Time.tsx
@@ -1,4 +1,4 @@
-import useStopwatchStore from '../../store/stopwatchStore';
+import useStopwatchStore from '../../store/stopwatch';
 import tw from 'tailwind-styled-components';
 import { formatTime } from '../../utils/stopwatch';
